+++ frontend/changelog.md
@@ -3,6 +3,9 @@
 **`race.py`**:
  - Added crash mechanics via two functions, 5% chance each lap a crash could happen.
 
+**`test_race.py`**:
+ - Added test for new crash mechanics
+
 # Dec 8 Changelog
 
 ## File Updates: